[SYCL] Part2: Split info_desc.hpp into per-trait-class headers
Split `<sycl/info/info_desc.hpp>` (~740 lines) into 6 per-trait-class
headers under `sycl/include/sycl/info/`:

* `platform.hpp` — 5 platform traits
* `context.hpp` — 7 context traits
* `device.hpp` — `device_type` / `partition_property` /
  `partition_affinity_domain` / `local_mem_type` / `fp_config` /
  `global_mem_cache_type` / `execution_capability` enums + ~95 device traits
* `queue.hpp` — 3 queue traits
* `kernel.hpp` — 5 kernel + 10 kernel_device_specific traits
* `event.hpp` — `event_command_status` enum + 2 event traits + 3
  event_profiling traits

`info_desc.hpp` becomes a thin umbrella that includes all six plus the
ext-trait chain (matrix, max_work_groups, forward_progress,
device_architecture, bindless_image_info, composite_device,
kernel_queue_info, ext::oneapi::info::device, ext::intel::info::{device,
kernel}, ext::codeplay::experimental::max_registers_query). Existing
consumers continue to work unchanged.

Prereq for trimming `info_desc.hpp` from internal consumers (next PR).
Per-class consumers of `info::queue` / `info::context` / etc. will be
able to skip `info::device`'s ~95 trait structs and the matrix /
device-architecture / forward-progress include chain.

This PR delivers no measurable compile-time win on its own (the
umbrella still pulls everything); it sets up the structural change.
